Move department view questions to own component
