+++ frontend/src/Shell/Shell.tsx
@@ -32,6 +32,7 @@ import { useThemeConstants } from '../Theme/mantineTheme';
 import classes from './Shell.module.css';
 import { HospitalView } from '../Components/Views/HospitalView/HospitalView';
 import { DepartmentView } from '../Components/Views/DepartmentView/DepartmentView';
+import { DepartmentViewQuestions } from '../Components/Views/DepartmentView/DepartmentViewQuestions';
 import { ProviderView } from '../Components/Views/ProviderView/ProviderView';
 import { SettingsView } from '../Components/Views/SettingsView/SettingsView';
 import { SelectedVisitsPanel } from '../Components/Toolbar/SelectedVisits/SelectedVisitsPanel';
@@ -373,6 +374,7 @@ export const Shell = observer(() => {
           {TABS.find((tab) => tab.key === store.state.ui.activeTab)?.content}
         </Container>
       </AppShell.Main>
+      {store.state.ui.activeTab === 'Department' && <DepartmentViewQuestions />}
 
       <Modal
         opened={aboutModalOpened}
